+++ rust/main/lander/src/adapter/chains/ethereum/nonce/state/assign/tests.rs
@@ -1,2 +1,2 @@
-mod assign;
-mod identify;
+mod assign_test;
+mod identify_test;
